+++ StRoot/StEStructPool/macros/doEStruct2ptDst.C
@@ -1,5 +1,5 @@
 /************************************************************************
- * $Id: doEStruct2ptDst.C,v 1.1 2003/10/15 18:20:57 porter Exp $
+ * $Id: doEStruct2ptDst.C,v 1.2 2004/04/15 18:46:32 msd Exp $
  *
  * Author: Jeff Porter 
  *
@@ -33,7 +33,7 @@ void doEStruct2ptDst(const char* fileListFile, const char* outputDir, const char
   // simple (global) centrality definition ...not persistant to event file.. 
   // and not used in this particular example
   StEStructCentrality* cent=StEStructCentrality::Instance();
-  const int temp[4]={0,4,7,50};
+  const double temp[4]={0,4,7,50};
   cent->setCentralities(temp,4);
 
   // create the low-level reader (here for EStructDst)
@@ -130,6 +130,9 @@ void doEStruct2ptDst(const char* fileListFile, const char* outputDir, const char
 /**********************************************************************
  *
  * $Log: doEStruct2ptDst.C,v $
+ * Revision 1.2  2004/04/15 18:46:32  msd
+ * Updated centrality variable types
+ *
  * Revision 1.1  2003/10/15 18:20:57  porter
  * initial check in of Estruct Analysis maker codes.
  *
